clean out and make edit profile layout function based
