+++ services/frontend-v3/src/pages/ProfileCreate.jsx
@@ -15,7 +15,6 @@ class ProfileCreate extends React.Component {
       <CreateProfileLayout
         changeLanguage={this.props.changeLanguage}
         keycloak={this.props.keycloak}
-        history={this.props.history}
         step={this.props.match.params.step}
       ></CreateProfileLayout>
     );
